requirements.txt
flask==1.0.0
flask-restful-0.3.7
flask-sqlalchemy-2.4.1
flask-migrate-2.5.2

